requirements.txt
kafka-python
pandas
scikit-learn
capymoa
streamlit
plotly
matplotlib
